requirements.txt
Flask==3.1.0
openai==1.14.3
twilio==9.5.2
python-dotenv==1.1.0
requests==2.32.3
schedule==1.2.2
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
